Adds extra_gem task
Allows the addition of arbitrary gems to be installed along side
manageiq-performance.  Can also be used with the `:solo_gem` task if you
wish to install multiple gems, but not `manageiq-performance`.

Note, currently this does not resolve dependencies and install those, so
only use this with gems with zero depedencies.