Allow spec paths to be used for extra/solo task
Allows using a local gem for installation in the extra_gem and solo_gem
installation tasks.

Does similar work to what is done for installing the
`manageiq-performance` gem in other tasks, where it reads the gemspec,
and builds it from that dir.  A few tweaks were necessary to then read
the gem name from the spec instead of from the arg, and generating the
tar_io when we are loading from a path (cache_file allows us to skip
this though, when it exists).